refactor: 검색 페이지의 로직과 UI 분리
